requirements.txt
Flask
Flask-HTTPAuth
Flask-SQLAlchemy
boto3
python-dotenv
fpdf
gunicorn
psycopg2-binary
Werkzeug